Serialization: Replace reflection-based serialization
The old solution was something I wrote up myself. Unfortunately, it
relied heavily on KReflect, which skirts around a lot of kotlinx-
serialization. This new solution, provided in part by "ephemient" on
the kotlinlang slack, handles everything through APIs provided by
kotlinx-serialization. It may be a bit more verbose in places
(requiring more annotations, namely), but it is a better design.