Bindings: Use SerenityOS/jakt repo; remove serde-based serialization
Because we are using our own type system instead of the Jakt compiler,
we no longer need to be able to serialize the entire project. The
external annotator is now only used to display error messages. This
lets us remove a lot of the complexity that surrounded serialization,
as well as removing the dependency on my custom fork.