Add a formatter
This is very incomplete, mostly because it was added for one primary
purpose: indenting the cursor when pressing after after a curly brace.
That feature alone makes the editing experience *so* much nicer. The
framework is there though, so it should be easy to add features, though
I'll probably wait until an "official" style guide for Jakt is made.