Grammar: Improve specialization parsing
Previously, the grammar had the classic less-than issue when parsing
generic specialization lists. This technically could have been fixed
by removing the "pin=1" on the GenericSpecialization rule, but doing
it "manually" through the ParserUtil allows a really nice error
message to be generated, and allows the lexer to fully recover.